patch 7.4.1976 Problem:    Number variables are not 64 bits while they could be. Solution:   Add the num64 feature. (Ken Takata)

+++ src/fileio.c
@@ -9481,7 +9481,7 @@ apply_autocmds_group(
 
 #ifdef FEAT_EVAL
 	/* set v:cmdarg (only when there is a matching pattern) */
-	save_cmdbang = get_vim_var_nr(VV_CMDBANG);
+	save_cmdbang = (long)get_vim_var_nr(VV_CMDBANG);
 	if (eap != NULL)
 	{
 	    save_cmdarg = set_cmdarg(eap, NULL);
